requirements.txt
pandas
GitPython
requests